refactor: abstracted helpers functions into separate module
